Fixes for deployment
