--- interfaces/feed.ts
@@ -1,6 +1,0 @@
-import { Item } from "./item";
-
-export type Feed = {
-    title: string,
-    items: Item[],
-}
