--- interfaces/tag.ts
@@ -1,6 +1,0 @@
-
-export type Tags = {[name: string]: Tag}
-
-export type Tag = {
-    name: string
-}
