+++ package.json
@@ -4,7 +4,7 @@
   "private": true,
   "scripts": {
     "dev": "NODE_OPTIONS='--inspect' next dev",
-    "build": "next build",
+    "build": "next build && next export",
     "start": "next start",
     "lint": "next lint"
   },
@@ -24,5 +24,8 @@
     "sass": "^1.57.1",
     "swr": "^2.0.0",
     "typescript": "4.9.4"
+  },
+  "devDependencies": {
+    "@types/lodash": "^4.14.191"
   }
 }
